transactional test

+++ was/src/main/java/com/example/drill/repository/TestRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TestRepository extends JpaRepository<User,Long> {
+public interface TestRepository extends JpaRepository<User,Integer> {
     List<User> findAll();
 
     User findByUserId(int userId);
